requirements.txt
httpx>=0.28,<1.0
google-genai>=1.0,<2.0
redis>=5.2,<6.0
apscheduler>=3.11,<4.0
python-dotenv>=1.1,<2.0
